create folder index.ts
