+++ src/app/pages/Cart/partials/CartList.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Image from './../../../../assets/images/index';
 import { ICart } from '../../../shared/interfaces/cart';
-import ButtonLink from '../../../shared/components/partials/Link';
+import { ButtonLink } from '../../../shared/components/partials/index';
 import CartItem from './CartItem';
 
 interface ICartListProps {
