+++ src/app/pages/Home/partials/SectionContact.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Button from '../../../shared/components/partials/Button';
+import { Button } from '../../../shared/components/partials/index';
 
 const SectionContact = () => {
   return (
