+++ src/app/pages/index.ts
@@ -1,0 +1,4 @@
+import Home from './Home/Home';
+import Cart from './Cart/Cart';
+
+export { Home, Cart };
