+++ src/app/shared/components/partials/Badge.tsx
@@ -4,9 +4,7 @@ interface IBadgeProps {
   discount: number;
   type: string;
 }
-
-const Badge: React.FC<IBadgeProps> = ({ discount, type }) => {
+ 
+export const Badge: React.FC<IBadgeProps> = ({ discount, type }) => {
   return <span className={type ? `badge badge-${type}`: "badge"}>-{discount}%</span>;
 };
-
-export default Badge;
