+++ src/app/shared/components/partials/Button.tsx
@@ -6,12 +6,10 @@ interface IButtonProps {
   onClick?: () => void
 }
 
-const Button: React.FC<IButtonProps> = ({ title, customClass, onClick }) => {
+export const Button: React.FC<IButtonProps> = ({ title, customClass, onClick }) => {
   return (
     <button className={customClass} onClick={onClick}>
       {title}
     </button>
   )
 }
-
-export default Button;
