+++ src/app/shared/components/partials/index.ts
@@ -1,0 +1,4 @@
+export { Badge } from './Badge';
+export { Button } from './Button';
+export { ButtonLink } from './Link';
+export { SectionTitle } from './SectionTitle';
